Enemy/bosspakkun perform: 88.3% -> 88.4% (retain signed phase guard)

+++ src/Enemy/bosspakkun.cpp
@@ -1219,7 +1219,7 @@ void TBossPakkun::perform(u32 flags, JDrama::TGraphics* graphics)
 			}
 		}
 
-		if (unk16C == 1 && checkMarioRiding()
+		if ((s8)unk16C == 1 && checkMarioRiding()
 		    && mSpine->getLatestNerve() != &TNerveBPJumpReact::theNerve()) {
 			mSpine->pushNerve(&TNerveBPJumpReact::theNerve());
 		}
